BIG CHANGES!!!
Scrapped project and rebuilt;

NiceGame is now using a Yume Nikki Fangame template, this makes things easier for us, but it also means that making changes to the system and customizing it to fit our needs will be much harder! Be careful! Try to understand the system and the format it follows!!